refactor: remove random + move popRandom to methods file
