+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/list/mutablelist/ext/MutableListExt.kt
@@ -3,7 +3,7 @@ package xyz.lbres.kotlinutils.list.mutablelist.ext
 /**
  * Remove a random element from list and return it
  *
- * @return [T]: an element from the list, or null if the list is empty
+ * @return [T]?: an element from the list, or `null` if the list is empty
  */
 fun <T> MutableList<T>.popRandom(): T? {
     if (isEmpty()) {
