+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/MutableMultiSet.kt
@@ -4,6 +4,4 @@ package xyz.lbres.kotlinutils.set.multiset
  * Interface for set that allows multiple occurrences of a value.
  * The interface supports read/write access to the values.
  */
-interface MutableMultiSet<E> : MultiSet<E>, MutableCollection<E> {
-    fun popRandom(): E?
-}
+interface MutableMultiSet<E> : MultiSet<E>, MutableCollection<E>
