+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MutableMultiSetMethodsTest.kt
@@ -1,0 +1,8 @@
+package xyz.lbres.kotlinutils.set.multiset
+
+import org.junit.Test
+import xyz.lbres.kotlinutils.set.multiset.methods.* // ktlint-disable no-wildcard-imports no-unused-imports
+
+class MutableMultiSetMethodsTest {
+    @Test fun testPopRandom() = runPopRandomTests()
+}
